+++ q1simulator/__init__.py
@@ -2,6 +2,6 @@ __version__ = "0.11.7"
 
 from .q1simulator import Q1Simulator
 from .cluster import Cluster, ClusterModule
-
+from .q1plotter import Q1Plotter
 from .q1viewer import plot_q1asm_file, plot_q1asm_files, PlotDef